requirements.txt
mysql-python
click==6.6
Flask==0.11.1
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.11.10
wheel==0.24.0
psycopg2
requests
xmltodict
flask-sqlalchemy
flask-mysql
Cython
numpy
